json fixes (product price now shows correct) + 2 config_2.json

+++ src/main/java/cz/cvut/fel/omo/smartfarm/model/products/Egg.java
@@ -1,7 +1,7 @@
 package cz.cvut.fel.omo.smartfarm.model.products;
 
 public class Egg extends Product {
-    public Egg(String name, int price, int weight) {
+    public Egg(String name, double price, int weight) {
         super(name, price, weight, ProductType.EGG);  // Передаем тип продукта
     }
 }
