+++ src/main/java/cz/cvut/fel/omo/smartfarm/model/products/Wool.java
@@ -1,7 +1,7 @@
 package cz.cvut.fel.omo.smartfarm.model.products;
 
 public class Wool extends Product {
-    public Wool(String name, int price, int weight) {
+    public Wool(String name, double price, int weight) {
         super(name, price, weight, ProductType.WOOL);
     }
 }
